move colorize_image to plotting tools
