+++ nuc_morph_analysis/analyses/density/watershed_validate.py
@@ -12,29 +12,6 @@ import matplotlib.cm as cm
 from nuc_morph_analysis.analyses.dataset_images_for_figures.figure_helper import return_glasbey_on_dark
 from skimage.measure import find_contours
 
-
-def colorize_image(mip, dft, feature='2d_area_nuc_cell_ratio'):
-    """
-    Function create an image where the segmentation image objects (e.g. nuclei)
-    are colored by the a given feature from the dataframe of that timepoint 
-
-    Parameters
-    ----------
-    mip : np.array
-        the max intensity projection of the labeled image
-    dft : pd.DataFrame
-        the dataframe of the timepoint
-    feature : str
-        the feature to color the image by
-    """
-    
-    # now recolor the image by matching the pixel values in image to label_img in dft
-    recolored_img = np.zeros_like(mip).astype('float32')
-    recolored_img[mip>0]=np.nan
-    for _,row in dft.iterrows():
-        recolored_img[mip==row['label_img']] = row[feature]
-    return recolored_img
-
 if __name__ == '__main__':
     # set the details
     TIMEPOINT = 48
