+++ nuc_morph_analysis/analyses/neighbor_of_X/run_all_example_worfklows.py
@@ -1,5 +1,5 @@
 #%%
-from nuc_morph_analysis.analyses.mitotic_neighbors import example_dying_track_neighbors
-from nuc_morph_analysis.analyses.mitotic_neighbors import example_mitotic_track_neighbors
-from nuc_morph_analysis.analyses.mitotic_neighbors import example_timepoint_neighbors_of_dying
-from nuc_morph_analysis.analyses.mitotic_neighbors import example_timepoint_of_neighbors_of_mitotic
+from nuc_morph_analysis.analyses.neighbor_of_X import example_dying_track_neighbors
+from nuc_morph_analysis.analyses.neighbor_of_X import example_mitotic_track_neighbors 
+from nuc_morph_analysis.analyses.neighbor_of_X import example_timepoint_neighbors_of_dying 
+from nuc_morph_analysis.analyses.neighbor_of_X import example_timepoint_of_neighbors_of_mitotic
